+++ route/proxy.go
@@ -1,6 +1,7 @@
 package route
 
 import (
+	"errors"
 	"log"
 	"net"
 	"net/http"
@@ -13,15 +14,17 @@ import (
 // Proxy is a dynamic reverse proxy.
 type Proxy struct {
 	tr             http.RoundTripper
+	localIP        string
 	clientIPHeader string
 	tlsHeader      string
 	tlsHeaderValue string
 	requests       gometrics.Timer
 }
 
-func NewProxy(tr http.RoundTripper, clientIPHeader, tlsHeader, tlsHeaderValue string) *Proxy {
+func NewProxy(tr http.RoundTripper, localIP, clientIPHeader, tlsHeader, tlsHeaderValue string) *Proxy {
 	return &Proxy{
 		tr:             tr,
+		localIP:        localIP,
 		clientIPHeader: clientIPHeader,
 		tlsHeader:      tlsHeader,
 		tlsHeaderValue: tlsHeaderValue,
@@ -42,17 +45,9 @@ func (p *Proxy) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	if p.clientIPHeader != "" {
-		ip, _, err := net.SplitHostPort(req.RemoteAddr)
-		if err != nil {
-			http.Error(w, "cannot parse "+req.RemoteAddr, http.StatusInternalServerError)
-			return
-		}
-		req.Header.Set(p.clientIPHeader, ip)
-	}
-
-	if p.tlsHeader != "" && req.TLS != nil {
-		req.Header.Set(p.tlsHeader, p.tlsHeaderValue)
+	if err := addHeaders(req, p.localIP, p.clientIPHeader, p.tlsHeader, p.tlsHeaderValue); err != nil {
+		http.Error(w, "cannot parse "+req.RemoteAddr, http.StatusInternalServerError)
+		return
 	}
 
 	start := time.Now()
@@ -61,4 +56,40 @@ func (p *Proxy) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	rp.ServeHTTP(w, req)
 	target.timer.UpdateSince(start)
 	p.requests.UpdateSince(start)
+}
+
+func addHeaders(r *http.Request, localIP, clientIPHeader, tlsHeader, tlsHeaderValue string) error {
+	remoteIP, _, err := net.SplitHostPort(r.RemoteAddr)
+	if err != nil {
+		return errors.New("cannot parse " + r.RemoteAddr)
+	}
+
+	if clientIPHeader != "" {
+		r.Header.Set(clientIPHeader, remoteIP)
+	}
+
+	xff := r.Header.Get("X-Forwarded-For")
+	if xff != "" && localIP != "" {
+		r.Header.Set("X-Forwarded-For", xff+", "+localIP)
+	}
+
+	fwd := r.Header.Get("Forwarded")
+	if fwd == "" {
+		fwd = "for=" + remoteIP
+		if r.TLS != nil {
+			fwd += "; proto=https"
+		} else {
+			fwd += "; proto=http"
+		}
+	}
+	if localIP != "" {
+		fwd += "; by=" + localIP
+	}
+	r.Header.Set("Forwarded", fwd)
+
+	if tlsHeader != "" && r.TLS != nil {
+		r.Header.Set(tlsHeader, tlsHeaderValue)
+	}
+
+	return nil
 }